requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1

# Anthropic
anthropic==0.42.0

# Google
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.159.0

# Supabase
supabase==2.11.0

# Async task queue (Redis)
arq==0.26.1

# Crypto
cryptography==44.0.0

# Security
python-jose[cryptography]==3.3.0
